requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-jwt-extended==4.5.3
pymongo==4.5.0
python-dotenv==1.0.0
apscheduler==3.10.1
boto3==1.28.55
spacy==3.6.1
scikit-learn==1.3.0
requests==2.31.0
werkzeug==2.3.7
gunicorn==21.2.0

# NLP model
# Install with: python -m spacy download en_core_web_sm 